Добавлен слой Presentantion и файлы для него.

+++ app/src/main/java/dev/passerby/aeon_project/domain/models/PaymentModel.kt
@@ -2,7 +2,7 @@ package dev.passerby.aeon_project.domain.models
 
 data class PaymentModel(
     val amount: Any?,
-    val created: Int?,
+    val created: Long?,
     val id: Int,
     val title: String
 )
